backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.2
django-cors-headers==4.6.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.11
packaging==26.0
requests==2.32.3
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.8.2
